Separate classes into different files.
