+++ lib/synvert/core/rewriter/condition/if_exist_condition.rb
@@ -1,0 +1,15 @@
+# encoding: utf-8
+
+module Synvert::Core
+  # IfExistCondition checks if matching node exists in the node children.
+  class Rewriter::IfExistCondition < Rewriter::Condition
+    # check if any child node matches the rules.
+    def match?
+      match = false
+      @instance.current_node.recursive_children do |child_node|
+        match = match || (child_node && child_node.match?(@rules))
+      end
+      match
+    end
+  end
+end
