+++ spec/synvert/core/rewriter/action/remove_action_spec.rb
@@ -1,0 +1,24 @@
+require 'spec_helper'
+
+module Synvert::Core
+  describe Rewriter::RemoveAction do
+    subject {
+      source = "user = User.new params[:user]\nuser.save\nrender\n"
+      send_node = Parser::CurrentRuby.parse(source).children[1]
+      instance = double(current_node: send_node)
+      Rewriter::RemoveAction.new(instance)
+    }
+
+    it 'gets begin_pos' do
+      expect(subject.begin_pos).to eq "user = User.new params[:user]\n".length
+    end
+
+    it 'gets end_pos' do
+      expect(subject.end_pos).to eq "user = User.new params[:user]\nuser.save".length
+    end
+
+    it 'gets rewritten_code' do
+      expect(subject.rewritten_code).to eq ""
+    end
+  end
+end
